+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebView.kt
@@ -160,7 +160,6 @@ internal open class MiniAppWebView(
     fun destroyView() {
         stopLoading()
         defaultFileDownloader.cleanup()
-        secureStorageDispatcher.cleanupSecureStorage()
         destroy()
     }